backend/requirements.txt
django>=6.0.3
djangorestframework
django-rest-framework-simplejwt
django-cors-headers
gunicorn
dj-database-url
psycopg2-binary
whitenoise
python-dotenv
